Switch to same font as used on conveyal.com
