+++ next.config.js
@@ -3,6 +3,9 @@ const withMDX = require('@zeit/next-mdx')({
 })
 
 module.exports = withMDX({
+  experimental: {
+    optimizeFonts: true
+  },
   pageExtensions: ['js', 'jsx', 'mdx', 'ts', 'tsx'],
   productionBrowserSourceMaps: true
 })
