+++ pages/_document.tsx
@@ -38,6 +38,10 @@ export default class extends Document {
     return (
       <Html>
         <Head>
+          <link
+            rel='stylesheet'
+            href='https://fonts.googleapis.com/css2?family=Inter:wght@400;600'
+          />
           <link rel='shortcut icon' href={LOGO_URL} type='image/x-icon' />
           <Stylesheets />
           {process.env.NEXT_PUBLIC_GA_TRACKING_ID && <Analytics />}
