+++ styles.css
@@ -14,8 +14,6 @@ html,
 body,
 #root,
 .tooltip {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-    Oxygen-Sans, Ubuntu, Cantarell, 'Helvetica Neue', sans-serif;
   font-size: 12px;
 }
